\name{predictedMeth}
\alias{predictedMeth}
\docType{data}
\title{
The output of \code{predictMeth}
}

\description{
  Please see the package vignette for description.
  }

\usage{data(predictedMeth)}
\format{
A \code{BSrel} object with the smoothed methylation data.
  }
\examples{
data(predictedMeth)
show(predictedMeth)
}
\keyword{datasets}
