\name{promoters}
\alias{promoters}
\docType{data}
\title{
A \code{GRanges} of promoters of the human genome
}

\description{
  Please see the package vignette for description.
  }

\usage{data(promoters)}
\format{
A \code{GRanges} object with the chromosomes, start and end positions of
defined human promoter regions together with an accesion number in the
\code{elementMetadata} slot.
}
\examples{
data(promoters)
head(promoters)
}
\keyword{datasets}
