\name{vario}
\alias{vario}
\docType{data}
\title{
Output of \code{makeVariogram}
}

\description{
  Please see the package vignette for description.
  }
\usage{data(vario)}
\format{
A \code{list} consisting of the \code{variogram} (a \code{matrix}) and
the \code{pValsList} (a \code{list} of the data frames of test results).
  }
\examples{
data(vario)
names(vario)
}
\keyword{datasets}

